+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Implement FUCA (Faire Un Choix Adéquat) 
 - Implement AROMAN (Alternative Ranking Order Method Accounting for Two-Step Normalization)
+- Implement GINI Coefficient Method for automatic weighting
 
 ### 0.8.4